# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhisperBox API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/create:
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a QRIS payment transaction for premium upgrade
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePaymentResponse"
        "400":
          description: Already premium or invalid request
        "401":
          description: Unauthorized
        "503":
          description: Payment service unavailable
  /payments/status/{ref}:
    get:
      operationId: getPaymentStatus
      tags: [payments]
      summary: Get payment transaction status
      security:
        - bearerAuth: []
      parameters:
        - name: ref
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatusResponse"
        "401":
          description: Unauthorized
        "404":
          description: Transaction not found
...
  /referrals/me:
    get:
      operationId: getMyReferralStats
      tags: [referrals]
      summary: Get referral stats and points
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Referral stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralStats"
        "401":
          description: Unauthorized
...
  /redeem:
    post:
      operationId: createRedeemRequest
      tags: [redeem]
      summary: Submit a point redeem request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRedeemRequestBody"
      responses:
        "201":
          description: Redeem request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRedeemRequestResponse"
        "400":
          description: Invalid request (not enough points or min not met)
        "401":
          description: Unauthorized
  /redeem/me:
    get:
      operationId: getMyRedeemRequests
      tags: [redeem]
      summary: Get my redeem requests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Redeem requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedeemRequestsResponse"
        "401":
          description: Unauthorized
  /referrals/claim:
    post:
      operationId: claimReferral
      tags: [referrals]
      summary: Claim a referral code after sign-up
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimReferralBody"
      responses:
        "200":
          description: Referral claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimReferralResponse"
        "400":
          description: Bad request (missing code or self-referral)
        "404":
          description: Referral code not found
        "409":
          description: Already claimed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    CreatePaymentResponse:
      type: object
      properties:
        merchantRef:
          type: string
        tripayRef:
          type: string
        amount:
          type: integer
        qrString:
          type: string
        qrImageUrl:
          type: string
        expiresAt:
          type: integer
      required:
        - merchantRef
        - amount
        - qrString
        - qrImageUrl
    PaymentStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [UNPAID, PAID, EXPIRED, FAILED]
        isPremium:
          type: boolean
        paidAt:
          type: string
          format: date-time
          nullable: true
      required:
        - status
        - isPremium
...
    ReferralUser:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
          nullable: true
        joinedAt:
          type: string
          format: date-time
        points:
          type: integer
        upgradeBonusAwarded:
          type: boolean
        isPremium:
          type: boolean
      required:
        - username
        - joinedAt
        - points
        - upgradeBonusAwarded
        - isPremium
    ReferralStats:
      type: object
      properties:
        referralCode:
          type: string
        points:
          type: integer
        pointsFromLinkOpens:
          type: integer
        totalPoints:
          type: integer
        availablePoints:
          type: integer
        redeemedPoints:
          type: integer
        linkOpens:
          type: integer
        referralCount:
          type: integer
        referrals:
          type: array
          items:
            $ref: "#/components/schemas/ReferralUser"
      required:
        - referralCode
        - points
        - pointsFromLinkOpens
        - totalPoints
        - availablePoints
        - redeemedPoints
        - linkOpens
        - referralCount
        - referrals
    ClaimReferralBody:
      type: object
      properties:
        referralCode:
          type: string
      required:
        - referralCode
    ClaimReferralResponse:
      type: object
      properties:
        success:
          type: boolean
        pointsAwarded:
          type: integer
      required:
        - success
        - pointsAwarded
...
    RedeemRequest:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        points:
          type: integer
        paymentInfo:
          type: string
        status:
          type: string
          enum: [pending, success, rejected]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - points
        - paymentInfo
        - status
        - createdAt
        - updatedAt
    RedeemRequestsResponse:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/RedeemRequest"
      required:
        - requests
    CreateRedeemRequestBody:
      type: object
      properties:
        points:
          type: integer
          minimum: 1000
        paymentInfo:
          type: string
          minLength: 1
      required:
        - points
        - paymentInfo
    CreateRedeemRequestResponse:
      type: object
      properties:
        request:
          $ref: "#/components/schemas/RedeemRequest"
      required:
        - request